Enabling to delete object even it doesn't have image file

+++ helpers.js
@@ -1,14 +1,17 @@
 const fs = require("fs");
 const path = require("path");
 
-const deleteFile = filename => {
-    fs.unlink(
-        path.join(__dirname, `./dist/images/uploads/${filename}`),
-        error => {
-            if (error) throw error;
-            console.log(`${filename} deleted!`);
-        }
-    );
+const deleteFile = file => {
+    const filePath = path.join(__dirname, "./dist/images/uploads/");
+    if (fs.existsSync(filePath + file)) {
+        fs.unlink(
+            path.join(__dirname, `./dist/images/uploads/${file}`),
+            error => {
+                if (error) throw error;
+                console.log(`${file} deleted!`);
+            }
+        );
+    }
 };
 
 const helpers = {
